Reduce RgbColor to 24 bits.

+++ include/ColorTable.hpp
@@ -34,11 +34,11 @@ class RgbColor {
     static const RgbColor kWhite;
 
     RgbColor();
-    RgbColor(uint16_t red, uint16_t green, uint16_t blue);
+    RgbColor(uint8_t red, uint8_t green, uint8_t blue);
 
-    uint16_t red;
-    uint16_t green;
-    uint16_t blue;
+    uint8_t red;
+    uint8_t green;
+    uint8_t blue;
 
     void read(BinaryReader* bin);
     void write(BinaryWriter* bin) const;
